Refactor error handling and parsing into modules
